requirements.txt.txt
torchvision=0.13
pytorch
tabulate
scipy
matplotlib
seaborn
torchinfo
pyhessian
termcolor